Struct rusoto_sagemaker::HyperParameterTrainingJobDefinition

pub struct HyperParameterTrainingJobDefinition {
    pub algorithm_specification: HyperParameterAlgorithmSpecification,
    pub input_data_config: Vec<Channel>,
    pub output_data_config: OutputDataConfig,
    pub resource_config: ResourceConfig,
    pub role_arn: String,
    pub static_hyper_parameters: Option<HashMap<String, String>>,
    pub stopping_condition: StoppingCondition,
    pub vpc_config: Option<VpcConfig>,
}

Defines the training jobs launched by a hyperparameter tuning job.

Fields
algorithm_specification: HyperParameterAlgorithmSpecification

The HyperParameterAlgorithmSpecification object that specifies the algorithm to use for the training jobs that the tuning job launches.

input_data_config: Vec<Channel>

An array of Channel objects that specify the input for the training jobs that the tuning job launches.

output_data_config: OutputDataConfig

Specifies the path to the Amazon S3 bucket where you store model artifacts from the training jobs that the tuning job launches.

resource_config: ResourceConfig

The resources, including the compute instances and storage volumes, to use for the training jobs that the tuning job launches.

Storage volumes store model artifacts and incremental states. Training algorithms might also use storage volumes for scratch space. If you want Amazon SageMaker to use the storage volume to store the training data, choose File as the TrainingInputMode in the algorithm specification. For distributed training algorithms, specify an instance count greater than 1.

role_arn: String

The Amazon Resource Name (ARN) of the IAM role associated with the training jobs that the tuning job launches.

static_hyper_parameters: Option<HashMap<String, String>>

Specifies the values of hyperparameters that do not change for the tuning job.

stopping_condition: StoppingCondition

Sets a maximum duration for the training jobs that the tuning job launches. Use this parameter to limit model training costs.

To stop a job, Amazon SageMaker sends the algorithm the SIGTERM signal. This delays job termination for 120 seconds. Algorithms might use this 120-second window to save the model artifacts.

When Amazon SageMaker terminates a job because the stopping condition has been met, training algorithms provided by Amazon SageMaker save the intermediate results of the job.

vpc_config: Option<VpcConfig>

The VpcConfig object that specifies the VPC that you want the training jobs that this hyperparameter tuning job launches to connect to. Control access to and from your training container by configuring the VPC. For more information, see train-vpc.