Struct rusoto_sagemaker::TrainingJob

pub struct TrainingJob {
    pub algorithm_specification: Option<AlgorithmSpecification>,
    pub creation_time: Option<f64>,
    pub enable_inter_container_traffic_encryption: Option<bool>,
    pub enable_network_isolation: Option<bool>,
    pub failure_reason: Option<String>,
    pub final_metric_data_list: Option<Vec<MetricData>>,
    pub hyper_parameters: Option<HashMap<String, String>>,
    pub input_data_config: Option<Vec<Channel>>,
    pub labeling_job_arn: Option<String>,
    pub last_modified_time: Option<f64>,
    pub model_artifacts: Option<ModelArtifacts>,
    pub output_data_config: Option<OutputDataConfig>,
    pub resource_config: Option<ResourceConfig>,
    pub role_arn: Option<String>,
    pub secondary_status: Option<String>,
    pub secondary_status_transitions: Option<Vec<SecondaryStatusTransition>>,
    pub stopping_condition: Option<StoppingCondition>,
    pub tags: Option<Vec<Tag>>,
    pub training_end_time: Option<f64>,
    pub training_job_arn: Option<String>,
    pub training_job_name: Option<String>,
    pub training_job_status: Option<String>,
    pub training_start_time: Option<f64>,
    pub tuning_job_arn: Option<String>,
    pub vpc_config: Option<VpcConfig>,
}

Contains information about a training job.

Fields
algorithm_specification: Option<AlgorithmSpecification>

Information about the algorithm used for training, and algorithm metadata.

creation_time: Option<f64>

A timestamp that indicates when the training job was created.

enable_inter_container_traffic_encryption: Option<bool>

To encrypt all communications between ML compute instances in distributed training, choose True. Encryption provides greater security for distributed training, but training might take longer. How long it takes depends on the amount of communication between compute instances, especially if you use a deep learning algorithm in distributed training.

enable_network_isolation: Option<bool>

If the TrainingJob was created with network isolation, the value is set to true. If network isolation is enabled, nodes can't communicate beyond the VPC they run in.

failure_reason: Option<String>

If the training job failed, the reason it failed.

final_metric_data_list: Option<Vec<MetricData>>

A list of final metric values that are set when the training job completes. Used only if the training job was configured to use metrics.

hyper_parameters: Option<HashMap<String, String>>

Algorithm-specific parameters.

input_data_config: Option<Vec<Channel>>

An array of Channel objects that describes each data input channel.

labeling_job_arn: Option<String>

The Amazon Resource Name (ARN) of the labeling job.

last_modified_time: Option<f64>

A timestamp that indicates when the status of the training job was last modified.

model_artifacts: Option<ModelArtifacts>

Information about the Amazon S3 location that is configured for storing model artifacts.

output_data_config: Option<OutputDataConfig>

The S3 path where model artifacts that you configured when creating the job are stored. Amazon SageMaker creates subfolders for model artifacts.

resource_config: Option<ResourceConfig>

Resources, including ML compute instances and ML storage volumes, that are configured for model training.

role_arn: Option<String>

The AWS Identity and Access Management (IAM) role configured for the training job.

secondary_status: Option<String>

Provides detailed information about the state of the training job. For detailed information about the secondary status of the training job, see StatusMessage under SecondaryStatusTransition.

Amazon SageMaker provides primary statuses and secondary statuses that apply to each of them:

InProgress

Starting - Starting the training job.

Downloading - An optional stage for algorithms that support File training input mode. It indicates that data is being downloaded to the ML storage volumes.

Training - Training is in progress.

Uploading - Training is complete and the model artifacts are being uploaded to the S3 location.

Completed

Completed - The training job has completed.

Failed

Failed - The training job has failed. The reason for the failure is returned in the FailureReason field of DescribeTrainingJobResponse.

Stopped

MaxRuntimeExceeded - The job stopped because it exceeded the maximum allowed runtime.

Stopped - The training job has stopped.

Stopping

Stopping - Stopping the training job.

Valid values for SecondaryStatus are subject to change.

We no longer support the following secondary statuses:

LaunchingMLInstances

PreparingTrainingStack

DownloadingTrainingImage

secondary_status_transitions: Option<Vec<SecondaryStatusTransition>>

A history of all of the secondary statuses that the training job has transitioned through.

stopping_condition: Option<StoppingCondition>

The condition under which to stop the training job.

tags: Option<Vec<Tag>>

An array of key-value pairs. For more information, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide.

training_end_time: Option<f64>

Indicates the time when the training job ends on training instances. You are billed for the time interval between the value of TrainingStartTime and this time. For successful jobs and stopped jobs, this is the time after model artifacts are uploaded. For failed jobs, this is the time when Amazon SageMaker detects a job failure.

training_job_arn: Option<String>

The Amazon Resource Name (ARN) of the training job.

training_job_name: Option<String>

The name of the training job.

training_job_status: Option<String>

The status of the training job.

Training job statuses are:

InProgress - The training is in progress.

Completed - The training job has completed.

Failed - The training job has failed. To see the reason for the failure, see the FailureReason field in the response to a DescribeTrainingJobResponse call.

Stopping - The training job is stopping.

Stopped - The training job has stopped.

For more detailed information, see SecondaryStatus.

training_start_time: Option<f64>

Indicates the time when the training job starts on training instances. You are billed for the time interval between this time and the value of TrainingEndTime. The start time in CloudWatch Logs might be later than this time. The difference is due to the time it takes to download the training data and to the size of the training container.

tuning_job_arn: Option<String>

The Amazon Resource Name (ARN) of the associated hyperparameter tuning job if the training job was launched by a hyperparameter tuning job.

vpc_config: Option<VpcConfig>

A VpcConfig object that specifies the VPC that this training job has access to. For more information, see Protect Training Jobs by Using an Amazon Virtual Private Cloud.