Crate rusoto_sagemaker

Provides APIs for creating and managing Amazon SageMaker resources.

If you're using the service, you're probably looking for SageMakerClient and SageMaker.

Structs

AddTagsInput

AddTagsOutput

AlgorithmSpecification
Specifies the training algorithm to use in a CreateTrainingJob request.

For more information about algorithms provided by Amazon SageMaker, see Algorithms. For information about using your own algorithms, see Using Your Own Algorithms with Amazon SageMaker.

AlgorithmStatusDetails
Specifies the validation and image scan statuses of the algorithm.

AlgorithmStatusItem
Represents the overall status of an algorithm.

AlgorithmSummary
Provides summary information about an algorithm.

AlgorithmValidationProfile
Defines a training job and a batch transform job that Amazon SageMaker runs to validate your algorithm.

The data provided in the validation profile is made available to your buyers on AWS Marketplace.

AlgorithmValidationSpecification
Specifies configurations for one or more training jobs that Amazon SageMaker runs to test the algorithm.

AnnotationConsolidationConfig
Configures how labels are consolidated across human workers.

CategoricalParameterRange
A list of categorical hyperparameters to tune.

CategoricalParameterRangeSpecification
Defines the possible values for a categorical hyperparameter.

Channel
A channel is a named input source that training algorithms can consume.

ChannelSpecification
Defines a named input source, called a channel, to be used by an algorithm.

CodeRepositorySummary
Specifies summary information about a Git repository.

CognitoMemberDefinition
Identifies a Amazon Cognito user group. A user group can be used in on or more work teams.

CompilationJobSummary
A summary of a model compilation job.

ContainerDefinition
Describes the container, as part of model definition.

ContinuousParameterRange
A list of continuous hyperparameters to tune.

ContinuousParameterRangeSpecification
Defines the possible values for a continuous hyperparameter.

CreateAlgorithmInput

CreateAlgorithmOutput

CreateCodeRepositoryInput

CreateCodeRepositoryOutput

CreateCompilationJobRequest

CreateCompilationJobResponse

CreateEndpointConfigInput

CreateEndpointConfigOutput

CreateEndpointInput

CreateEndpointOutput

CreateHyperParameterTuningJobRequest

CreateHyperParameterTuningJobResponse

CreateLabelingJobRequest

CreateLabelingJobResponse

CreateModelInput

CreateModelOutput

CreateModelPackageInput

CreateModelPackageOutput

CreateNotebookInstanceInput

CreateNotebookInstanceLifecycleConfigInput

CreateNotebookInstanceLifecycleConfigOutput

CreateNotebookInstanceOutput

CreatePresignedNotebookInstanceUrlInput

CreatePresignedNotebookInstanceUrlOutput

CreateTrainingJobRequest

CreateTrainingJobResponse

CreateTransformJobRequest

CreateTransformJobResponse

CreateWorkteamRequest

CreateWorkteamResponse

DataSource
Describes the location of the channel data.

DeleteAlgorithmInput

DeleteCodeRepositoryInput

DeleteEndpointConfigInput

DeleteEndpointInput

DeleteModelInput

DeleteModelPackageInput

DeleteNotebookInstanceInput

DeleteNotebookInstanceLifecycleConfigInput

DeleteTagsInput

DeleteTagsOutput

DeleteWorkteamRequest

DeleteWorkteamResponse

DeployedImage
Gets the Amazon EC2 Container Registry path of the docker image of the model that is hosted in this ProductionVariant.

If you used the registry/repository[:tag] form to specify the image path of the primary container when you created the model hosted in this ProductionVariant, the path resolves to a path of the form registry/repository[@digest]. A digest is a hash value that identifies a specific version of an image. For information about Amazon ECR paths, see Pulling an Image in the Amazon ECR User Guide.

DescribeAlgorithmInput

DescribeAlgorithmOutput

DescribeCodeRepositoryInput

DescribeCodeRepositoryOutput

DescribeCompilationJobRequest

DescribeCompilationJobResponse

DescribeEndpointConfigInput

DescribeEndpointConfigOutput

DescribeEndpointInput

DescribeEndpointOutput

DescribeHyperParameterTuningJobRequest

DescribeHyperParameterTuningJobResponse

DescribeLabelingJobRequest

DescribeLabelingJobResponse

DescribeModelInput

DescribeModelOutput

DescribeModelPackageInput

DescribeModelPackageOutput

DescribeNotebookInstanceInput

DescribeNotebookInstanceLifecycleConfigInput

DescribeNotebookInstanceLifecycleConfigOutput

DescribeNotebookInstanceOutput

DescribeSubscribedWorkteamRequest

DescribeSubscribedWorkteamResponse

DescribeTrainingJobRequest

DescribeTrainingJobResponse

DescribeTransformJobRequest

DescribeTransformJobResponse

DescribeWorkteamRequest

DescribeWorkteamResponse

DesiredWeightAndCapacity
Specifies weight and capacity values for a production variant.

EndpointConfigSummary
Provides summary information for an endpoint configuration.

EndpointSummary
Provides summary information for an endpoint.

Filter

A conditional statement for a search expression that includes a Boolean operator, a resource property, and a value.

If you don't specify an Operator and a Value, the filter searches for only the specified property. For example, defining a Filter for the FailureReason for the TrainingJob Resource searches for training job objects that have a value in the FailureReason field.

If you specify a Value, but not an Operator, Amazon SageMaker uses the equals operator as the default.

In search, there are several property types:

Metrics

To define a metric filter, enter a value using the form "Metrics.<name>", where <name> is a metric name. For example, the following filter searches for training jobs with an "accuracy" metric greater than "0.9":

{

"Name": "Metrics.accuracy",

"Operator": "GREATERTHAN",

"Value": "0.9"

}

HyperParameters

To define a hyperparameter filter, enter a value with the form "HyperParameters.<name>". Decimal hyperparameter values are treated as a decimal in a comparison if the specified Value is also a decimal value. If the specified Value is an integer, the decimal hyperparameter values are treated as integers. For example, the following filter is satisfied by training jobs with a "learningrate" hyperparameter that is less than "0.5":

{

"Name": "HyperParameters.learningrate",

"Operator": "LESSTHAN",

"Value": "0.5"

}

Tags

To define a tag filter, enter a value with the form "Tags.<key>".

FinalHyperParameterTuningJobObjectiveMetric
Shows the final value for the objective metric for a training job that was launched by a hyperparameter tuning job. You define the objective metric in the HyperParameterTuningJobObjective parameter of HyperParameterTuningJobConfig.

GetSearchSuggestionsRequest

GetSearchSuggestionsResponse

GitConfig
Specifies configuration details for a Git repository in your AWS account.

GitConfigForUpdate
Specifies configuration details for a Git repository when the repository is updated.

HumanTaskConfig
Information required for human workers to complete a labeling task.

HyperParameterAlgorithmSpecification
Specifies which training algorithm to use for training jobs that a hyperparameter tuning job launches and the metrics to monitor.

HyperParameterSpecification
Defines a hyperparameter to be used by an algorithm.

HyperParameterTrainingJobDefinition
Defines the training jobs launched by a hyperparameter tuning job.

HyperParameterTrainingJobSummary
Specifies summary information about a training job.

HyperParameterTuningJobConfig
Configures a hyperparameter tuning job.

HyperParameterTuningJobObjective
Defines the objective metric for a hyperparameter tuning job. Hyperparameter tuning uses the value of this metric to evaluate the training jobs it launches, and returns the training job that results in either the highest or lowest value for this metric, depending on the value you specify for the Type parameter.

HyperParameterTuningJobSummary
Provides summary information about a hyperparameter tuning job.

HyperParameterTuningJobWarmStartConfig

Specifies the configuration for a hyperparameter tuning job that uses one or more previous hyperparameter tuning jobs as a starting point. The results of previous tuning jobs are used to inform which combinations of hyperparameters to search over in the new tuning job.

All training jobs launched by the new hyperparameter tuning job are evaluated by using the objective metric, and the training job that performs the best is compared to the best training jobs from the parent tuning jobs. From these, the training job that performs the best as measured by the objective metric is returned as the overall best training job.

All training jobs launched by parent hyperparameter tuning jobs and the new hyperparameter tuning jobs count against the limit of training jobs for the tuning job.

InferenceSpecification
Defines how to perform inference generation after a training job is run.

InputConfig
Contains information about the location of input model artifacts, the name and shape of the expected data inputs, and the framework in which the model was trained.

IntegerParameterRange
For a hyperparameter of the integer type, specifies the range that a hyperparameter tuning job searches.

IntegerParameterRangeSpecification
Defines the possible values for an integer hyperparameter.

LabelCounters
Provides a breakdown of the number of objects labeled.

LabelCountersForWorkteam
Provides counts for human-labeled tasks in the labeling job.

LabelingJobAlgorithmsConfig
Provides configuration information for auto-labeling of your data objects. A LabelingJobAlgorithmsConfig object must be supplied in order to use auto-labeling.

LabelingJobDataAttributes
Attributes of the data specified by the customer. Use these to describe the data to be labeled.

LabelingJobDataSource
Provides information about the location of input data.

LabelingJobForWorkteamSummary
Provides summary information for a work team.

LabelingJobInputConfig
Input configuration information for a labeling job.

LabelingJobOutput
Specifies the location of the output produced by the labeling job.

LabelingJobOutputConfig
Output configuration information for a labeling job.

LabelingJobResourceConfig
Provides configuration information for labeling jobs.

LabelingJobS3DataSource
The Amazon S3 location of the input data objects.

LabelingJobStoppingConditions
A set of conditions for stopping a labeling job. If any of the conditions are met, the job is automatically stopped. You can use these conditions to control the cost of data labeling.

LabelingJobSummary
Provides summary information about a labeling job.

ListAlgorithmsInput

ListAlgorithmsOutput

ListCodeRepositoriesInput

ListCodeRepositoriesOutput

ListCompilationJobsRequest

ListCompilationJobsResponse

ListEndpointConfigsInput

ListEndpointConfigsOutput

ListEndpointsInput

ListEndpointsOutput

ListHyperParameterTuningJobsRequest

ListHyperParameterTuningJobsResponse

ListLabelingJobsForWorkteamRequest

ListLabelingJobsForWorkteamResponse

ListLabelingJobsRequest

ListLabelingJobsResponse

ListModelPackagesInput

ListModelPackagesOutput

ListModelsInput

ListModelsOutput

ListNotebookInstanceLifecycleConfigsInput

ListNotebookInstanceLifecycleConfigsOutput

ListNotebookInstancesInput

ListNotebookInstancesOutput

ListSubscribedWorkteamsRequest

ListSubscribedWorkteamsResponse

ListTagsInput

ListTagsOutput

ListTrainingJobsForHyperParameterTuningJobRequest

ListTrainingJobsForHyperParameterTuningJobResponse

ListTrainingJobsRequest

ListTrainingJobsResponse

ListTransformJobsRequest

ListTransformJobsResponse

ListWorkteamsRequest

ListWorkteamsResponse

MemberDefinition
Defines the Amazon Cognito user group that is part of a work team.

MetricData
The name, value, and date and time of a metric that was emitted to Amazon CloudWatch.

MetricDefinition
Specifies a metric that the training algorithm writes to stderr or stdout. Amazon SageMakerhyperparameter tuning captures all defined metrics. You specify one metric that a hyperparameter tuning job uses as its objective metric to choose the best training job.

ModelArtifacts
Provides information about the location that is configured for storing model artifacts.

ModelPackageContainerDefinition
Describes the Docker container for the model package.

ModelPackageStatusDetails
Specifies the validation and image scan statuses of the model package.

ModelPackageStatusItem
Represents the overall status of a model package.

ModelPackageSummary
Provides summary information about a model package.

ModelPackageValidationProfile
Contains data, such as the inputs and targeted instance types that are used in the process of validating the model package.

The data provided in the validation profile is made available to your buyers on AWS Marketplace.

ModelPackageValidationSpecification
Specifies batch transform jobs that Amazon SageMaker runs to validate your model package.

ModelSummary
Provides summary information about a model.

NestedFilters

Defines a list of NestedFilters objects. To satisfy the conditions specified in the NestedFilters call, a resource must satisfy the conditions of all of the filters.

For example, you could define a NestedFilters using the training job's InputDataConfig property to filter on Channel objects.

A NestedFilters object contains multiple filters. For example, to find all training jobs whose name contains train and that have cat/data in their S3Uri (specified in InputDataConfig), you need to create a NestedFilters object that specifies the InputDataConfig property with the following Filter objects:

'{Name:"InputDataConfig.ChannelName", "Operator":"EQUALS", "Value":"train"}',

'{Name:"InputDataConfig.DataSource.S3DataSource.S3Uri", "Operator":"CONTAINS", "Value":"cat/data"}'

NotebookInstanceLifecycleConfigSummary
Provides a summary of a notebook instance lifecycle configuration.

NotebookInstanceLifecycleHook
Contains the notebook instance lifecycle configuration script.

Each lifecycle configuration script has a limit of 16384 characters.

The value of the $PATH environment variable that is available to both scripts is /sbin:bin:/usr/sbin:/usr/bin.

View CloudWatch Logs for notebook instance lifecycle configurations in log group /aws/sagemaker/NotebookInstances in log stream [notebook-instance-name]/[LifecycleConfigHook].

Lifecycle configuration scripts cannot run for longer than 5 minutes. If a script runs for longer than 5 minutes, it fails and the notebook instance is not created or started.

For information about notebook instance lifestyle configurations, see Step 2.1: (Optional) Customize a Notebook Instance.

NotebookInstanceSummary
Provides summary information for an Amazon SageMaker notebook instance.

NotificationConfiguration
Configures SNS notifications of available or expiring work items for work teams.

ObjectiveStatusCounters
Specifies the number of training jobs that this hyperparameter tuning job launched, categorized by the status of their objective metric. The objective metric status shows whether the final objective metric for the training job has been evaluated by the tuning job and used in the hyperparameter tuning process.

OutputConfig
Contains information about the output location for the compiled model and the device (target) that the model runs on.

OutputDataConfig
Provides information about how to store model training results (model artifacts).

ParameterRange
Defines the possible values for categorical, continuous, and integer hyperparameters to be used by an algorithm.

ParameterRanges

Specifies ranges of integer, continuous, and categorical hyperparameters that a hyperparameter tuning job searches. The hyperparameter tuning job launches training jobs with hyperparameter values within these ranges to find the combination of values that result in the training job with the best performance as measured by the objective metric of the hyperparameter tuning job.

You can specify a maximum of 20 hyperparameters that a hyperparameter tuning job can search over. Every possible value of a categorical parameter range counts against this limit.

ParentHyperParameterTuningJob
A previously completed or stopped hyperparameter tuning job to be used as a starting point for a new hyperparameter tuning job.

ProductionVariant
Identifies a model that you want to host and the resources to deploy for hosting it. If you are deploying multiple models, tell Amazon SageMaker how to distribute traffic among the models by specifying variant weights.

ProductionVariantSummary
Describes weight and capacities for a production variant associated with an endpoint. If you sent a request to the UpdateEndpointWeightsAndCapacities API and the endpoint status is Updating, you get different desired and current values.

PropertyNameQuery
A type of SuggestionQuery. A suggestion query for retrieving property names that match the specified hint.

PropertyNameSuggestion
A property name returned from a GetSearchSuggestions call that specifies a value in the PropertyNameQuery field.

PublicWorkforceTaskPrice

Defines the amount of money paid to an Amazon Mechanical Turk worker for each task performed.

Use one of the following prices for bounding box tasks. Prices are in US dollars.

0.036

0.048

0.060

0.072

0.120

0.240

0.360

0.480

0.600

0.720

0.840

0.960

1.080

1.200

Use one of the following prices for image classification, text classification, and custom tasks. Prices are in US dollars.

0.012

0.024

0.036

0.048

0.060

0.072

0.120

0.240

0.360

0.480

0.600

0.720

0.840

0.960

1.080

1.200

Use one of the following prices for semantic segmentation tasks. Prices are in US dollars.

0.840

0.960

1.080

1.200

RenderUiTemplateRequest

RenderUiTemplateResponse

RenderableTask
Contains input values for a task.

RenderingError
A description of an error that occurred while rendering the template.

ResourceConfig
Describes the resources, including ML compute instances and ML storage volumes, to use for model training.

ResourceLimits
Specifies the maximum number of training jobs and parallel training jobs that a hyperparameter tuning job can launch.

S3DataSource
Describes the S3 data source.

SageMakerClient
A client for the SageMaker API.

SearchExpression

A multi-expression that searches for the specified resource or resources in a search. All resource objects that satisfy the expression's condition are included in the search results. You must specify at least one subexpression, filter, or nested filter. A SearchExpression can contain up to twenty elements.

A SearchExpression contains the following components:

A list of Filter objects. Each filter defines a simple Boolean expression comprised of a resource property name, Boolean operator, and value.

A list of NestedFilter objects. Each nested filter defines a list of Boolean expressions using a list of resource properties. A nested filter is satisfied if a single object in the list satisfies all Boolean expressions.

A list of SearchExpression objects. A search expression object can be nested in a list of search expression objects.

A Boolean operator: And or Or.

SearchRecord
An individual search result record that contains a single resource object.

SearchRequest

SearchResponse

SecondaryStatusTransition

An array element of DescribeTrainingJobResponse$SecondaryStatusTransitions. It provides additional details about a status that the training job has transitioned through. A training job can be in one of several states, for example, starting, downloading, training, or uploading. Within each state, there are a number of intermediate states. For example, within the starting state, Amazon SageMaker could be starting the training job or launching the ML instances. These transitional states are referred to as the job's secondary status.

ShuffleConfig
A configuration for a shuffle option for input data in a channel. If you use S3Prefix for S3DataType, the results of the S3 key prefix matches are shuffled. If you use ManifestFile, the order of the S3 object references in the ManifestFile is shuffled. If you use AugmentedManifestFile, the order of the JSON lines in the AugmentedManifestFile is shuffled. The shuffling order is determined using the Seed value.

For Pipe input mode, shuffling is done at the start of every epoch. With large datasets, this ensures that the order of the training data is different for each epoch, and it helps reduce bias and possible overfitting. In a multi-node training job when ShuffleConfig is combined with S3DataDistributionType of ShardedByS3Key, the data is shuffled across nodes so that the content sent to a particular node on the first epoch might be sent to a different node on the second epoch.

SourceAlgorithm
Specifies an algorithm that was used to create the model package. The algorithm must be either an algorithm resource in your Amazon SageMaker account or an algorithm in AWS Marketplace that you are subscribed to.

SourceAlgorithmSpecification
A list of algorithms that were used to create a model package.

StartNotebookInstanceInput

StopCompilationJobRequest

StopHyperParameterTuningJobRequest

StopLabelingJobRequest

StopNotebookInstanceInput

StopTrainingJobRequest

StopTransformJobRequest

StoppingCondition
Specifies how long model training can run. When model training reaches the limit, Amazon SageMaker ends the training job. Use this API to cap model training cost.

To stop a job, Amazon SageMaker sends the algorithm the SIGTERM signal, which delays job termination for120 seconds. Algorithms might use this 120-second window to save the model artifacts, so the results of training is not lost.

Training algorithms provided by Amazon SageMaker automatically saves the intermediate results of a model training job (it is best effort case, as model might not be ready to save as some stages, for example training just started). This intermediate data is a valid model artifact. You can use it to create a model (CreateModel).

SubscribedWorkteam
Describes a work team of a vendor that does the a labelling job.

SuggestionQuery
Limits the property names that are included in the response.

Tag
Describes a tag.

TrainingJob
Contains information about a training job.

TrainingJobDefinition
Defines the input needed to run a training job using the algorithm.

TrainingJobStatusCounters
The numbers of training jobs launched by a hyperparameter tuning job, categorized by status.

TrainingJobSummary
Provides summary information about a training job.

TrainingSpecification
Defines how the algorithm is used for a training job.

TransformDataSource
Describes the location of the channel data.

TransformInput
Describes the input source of a transform job and the way the transform job consumes it.

TransformJobDefinition
Defines the input needed to run a transform job using the inference specification specified in the algorithm.

TransformJobSummary
Provides a summary of a transform job. Multiple TransformJobSummary objects are returned as a list after in response to a ListTransformJobs call.

TransformOutput
Describes the results of a transform job.

TransformResources
Describes the resources, including ML instance types and ML instance count, to use for transform job.

TransformS3DataSource
Describes the S3 data source.

USD
Represents an amount of money in United States dollars/

UiConfig
Provided configuration information for the worker UI for a labeling job.

UiTemplate
The Liquid template for the worker user interface.

UpdateCodeRepositoryInput

UpdateCodeRepositoryOutput

UpdateEndpointInput

UpdateEndpointOutput

UpdateEndpointWeightsAndCapacitiesInput

UpdateEndpointWeightsAndCapacitiesOutput

UpdateNotebookInstanceInput

UpdateNotebookInstanceLifecycleConfigInput

UpdateNotebookInstanceLifecycleConfigOutput

UpdateNotebookInstanceOutput

UpdateWorkteamRequest

UpdateWorkteamResponse

VpcConfig
Specifies a VPC that your training jobs and hosted models have access to. Control access to and from your training and model containers by configuring the VPC. For more information, see Protect Endpoints by Using an Amazon Virtual Private Cloud and Protect Training Jobs by Using an Amazon Virtual Private Cloud.

Workteam
Provides details about a labeling work team.

Enums

AddTagsError
Errors returned by AddTags

CreateAlgorithmError
Errors returned by CreateAlgorithm

CreateCodeRepositoryError
Errors returned by CreateCodeRepository

CreateCompilationJobError
Errors returned by CreateCompilationJob

CreateEndpointConfigError
Errors returned by CreateEndpointConfig

CreateEndpointError
Errors returned by CreateEndpoint

CreateHyperParameterTuningJobError
Errors returned by CreateHyperParameterTuningJob

CreateLabelingJobError
Errors returned by CreateLabelingJob

CreateModelError
Errors returned by CreateModel

CreateModelPackageError
Errors returned by CreateModelPackage

CreateNotebookInstanceError
Errors returned by CreateNotebookInstance

CreateNotebookInstanceLifecycleConfigError
Errors returned by CreateNotebookInstanceLifecycleConfig

CreatePresignedNotebookInstanceUrlError
Errors returned by CreatePresignedNotebookInstanceUrl

CreateTrainingJobError
Errors returned by CreateTrainingJob

CreateTransformJobError
Errors returned by CreateTransformJob

CreateWorkteamError
Errors returned by CreateWorkteam

DeleteAlgorithmError
Errors returned by DeleteAlgorithm

DeleteCodeRepositoryError
Errors returned by DeleteCodeRepository

DeleteEndpointConfigError
Errors returned by DeleteEndpointConfig

DeleteEndpointError
Errors returned by DeleteEndpoint

DeleteModelError
Errors returned by DeleteModel

DeleteModelPackageError
Errors returned by DeleteModelPackage

DeleteNotebookInstanceError
Errors returned by DeleteNotebookInstance

DeleteNotebookInstanceLifecycleConfigError
Errors returned by DeleteNotebookInstanceLifecycleConfig

DeleteTagsError
Errors returned by DeleteTags

DeleteWorkteamError
Errors returned by DeleteWorkteam

DescribeAlgorithmError
Errors returned by DescribeAlgorithm

DescribeCodeRepositoryError
Errors returned by DescribeCodeRepository

DescribeCompilationJobError
Errors returned by DescribeCompilationJob

DescribeEndpointConfigError
Errors returned by DescribeEndpointConfig

DescribeEndpointError
Errors returned by DescribeEndpoint

DescribeHyperParameterTuningJobError
Errors returned by DescribeHyperParameterTuningJob

DescribeLabelingJobError
Errors returned by DescribeLabelingJob

DescribeModelError
Errors returned by DescribeModel

DescribeModelPackageError
Errors returned by DescribeModelPackage

DescribeNotebookInstanceError
Errors returned by DescribeNotebookInstance

DescribeNotebookInstanceLifecycleConfigError
Errors returned by DescribeNotebookInstanceLifecycleConfig

DescribeSubscribedWorkteamError
Errors returned by DescribeSubscribedWorkteam

DescribeTrainingJobError
Errors returned by DescribeTrainingJob

DescribeTransformJobError
Errors returned by DescribeTransformJob

DescribeWorkteamError
Errors returned by DescribeWorkteam

GetSearchSuggestionsError
Errors returned by GetSearchSuggestions

ListAlgorithmsError
Errors returned by ListAlgorithms

ListCodeRepositoriesError
Errors returned by ListCodeRepositories

ListCompilationJobsError
Errors returned by ListCompilationJobs

ListEndpointConfigsError
Errors returned by ListEndpointConfigs

ListEndpointsError
Errors returned by ListEndpoints

ListHyperParameterTuningJobsError
Errors returned by ListHyperParameterTuningJobs

ListLabelingJobsError
Errors returned by ListLabelingJobs

ListLabelingJobsForWorkteamError
Errors returned by ListLabelingJobsForWorkteam

ListModelPackagesError
Errors returned by ListModelPackages

ListModelsError
Errors returned by ListModels

ListNotebookInstanceLifecycleConfigsError
Errors returned by ListNotebookInstanceLifecycleConfigs

ListNotebookInstancesError
Errors returned by ListNotebookInstances

ListSubscribedWorkteamsError
Errors returned by ListSubscribedWorkteams

ListTagsError
Errors returned by ListTags

ListTrainingJobsError
Errors returned by ListTrainingJobs

ListTrainingJobsForHyperParameterTuningJobError
Errors returned by ListTrainingJobsForHyperParameterTuningJob

ListTransformJobsError
Errors returned by ListTransformJobs

ListWorkteamsError
Errors returned by ListWorkteams

RenderUiTemplateError
Errors returned by RenderUiTemplate

SearchError
Errors returned by Search

StartNotebookInstanceError
Errors returned by StartNotebookInstance

StopCompilationJobError
Errors returned by StopCompilationJob

StopHyperParameterTuningJobError
Errors returned by StopHyperParameterTuningJob

StopLabelingJobError
Errors returned by StopLabelingJob

StopNotebookInstanceError
Errors returned by StopNotebookInstance

StopTrainingJobError
Errors returned by StopTrainingJob

StopTransformJobError
Errors returned by StopTransformJob

UpdateCodeRepositoryError
Errors returned by UpdateCodeRepository

UpdateEndpointError
Errors returned by UpdateEndpoint

UpdateEndpointWeightsAndCapacitiesError
Errors returned by UpdateEndpointWeightsAndCapacities

UpdateNotebookInstanceError
Errors returned by UpdateNotebookInstance

UpdateNotebookInstanceLifecycleConfigError
Errors returned by UpdateNotebookInstanceLifecycleConfig

UpdateWorkteamError
Errors returned by UpdateWorkteam

Traits

SageMaker
Trait representing the capabilities of the SageMaker API. SageMaker clients implement this trait.