Crate rusoto_sagemaker

Definition of the public APIs exposed by SageMaker

If you're using the service, you're probably looking for SageMakerClient and SageMaker.

Structs

AddTagsInput

AddTagsOutput

AlgorithmSpecification
Specifies the training algorithm to use in a CreateTrainingJob request.

For more information about algorithms provided by Amazon SageMaker, see Algorithms. For information about using your own algorithms, see your-algorithms.

CategoricalParameterRange
A list of categorical hyperparameters to tune.

Channel
A channel is a named input source that training algorithms can consume.

ContainerDefinition
Describes the container, as part of model definition.

ContinuousParameterRange
A list of continuous hyperparameters to tune.

CreateEndpointConfigInput

CreateEndpointConfigOutput

CreateEndpointInput

CreateEndpointOutput

CreateHyperParameterTuningJobRequest

CreateHyperParameterTuningJobResponse

CreateModelInput

CreateModelOutput

CreateNotebookInstanceInput

CreateNotebookInstanceLifecycleConfigInput

CreateNotebookInstanceLifecycleConfigOutput

CreateNotebookInstanceOutput

CreatePresignedNotebookInstanceUrlInput

CreatePresignedNotebookInstanceUrlOutput

CreateTrainingJobRequest

CreateTrainingJobResponse

CreateTransformJobRequest

CreateTransformJobResponse

DataSource
Describes the location of the channel data.

DeleteEndpointConfigInput

DeleteEndpointInput

DeleteModelInput

DeleteNotebookInstanceInput

DeleteNotebookInstanceLifecycleConfigInput

DeleteTagsInput

DeleteTagsOutput

DeployedImage
Gets the Amazon EC2 Container Registry path of the docker image of the model that is hosted in this ProductionVariant.

If you used the registry/repository[:tag] form to to specify the image path of the primary container when you created the model hosted in this ProductionVariant, the path resolves to a path of the form registry/repository[@digest]. A digest is a hash value that identifies a specific version of an image. For information about Amazon ECR paths, see Pulling an Image in the Amazon ECR User Guide.

DescribeEndpointConfigInput

DescribeEndpointConfigOutput

DescribeEndpointInput

DescribeEndpointOutput

DescribeHyperParameterTuningJobRequest

DescribeHyperParameterTuningJobResponse

DescribeModelInput

DescribeModelOutput

DescribeNotebookInstanceInput

DescribeNotebookInstanceLifecycleConfigInput

DescribeNotebookInstanceLifecycleConfigOutput

DescribeNotebookInstanceOutput

DescribeTrainingJobRequest

DescribeTrainingJobResponse

DescribeTransformJobRequest

DescribeTransformJobResponse

DesiredWeightAndCapacity
Specifies weight and capacity values for a production variant.

EndpointConfigSummary
Provides summary information for an endpoint configuration.

EndpointSummary
Provides summary information for an endpoint.

FinalHyperParameterTuningJobObjectiveMetric
Shows the final value for the objective metric for a training job that was launched by a hyperparameter tuning job. You define the objective metric in the HyperParameterTuningJobObjective parameter of HyperParameterTuningJobConfig.

HyperParameterAlgorithmSpecification
Specifies which training algorithm to use for training jobs that a hyperparameter tuning job launches and the metrics to monitor.

HyperParameterTrainingJobDefinition
Defines the training jobs launched by a hyperparameter tuning job.

HyperParameterTrainingJobSummary
Specifies summary information about a training job.

HyperParameterTuningJobConfig
Configures a hyperparameter tuning job.

HyperParameterTuningJobObjective
Defines the objective metric for a hyperparameter tuning job. Hyperparameter tuning uses the value of this metric to evaluate the training jobs it launches, and returns the training job that results in either the highest or lowest value for this metric, depending on the value you specify for the Type parameter.

HyperParameterTuningJobSummary
Provides summary information about a hyperparameter tuning job.

IntegerParameterRange
For a hyperparameter of the integer type, specifies the range that a hyperparameter tuning job searches.

ListEndpointConfigsInput

ListEndpointConfigsOutput

ListEndpointsInput

ListEndpointsOutput

ListHyperParameterTuningJobsRequest

ListHyperParameterTuningJobsResponse

ListModelsInput

ListModelsOutput

ListNotebookInstanceLifecycleConfigsInput

ListNotebookInstanceLifecycleConfigsOutput

ListNotebookInstancesInput

ListNotebookInstancesOutput

ListTagsInput

ListTagsOutput

ListTrainingJobsForHyperParameterTuningJobRequest

ListTrainingJobsForHyperParameterTuningJobResponse

ListTrainingJobsRequest

ListTrainingJobsResponse

ListTransformJobsRequest

ListTransformJobsResponse

MetricDefinition
Specifies a metric that the training algorithm writes to stderr or stdout. Amazon SageMakerHyperparamter tuning captures all defined metrics. You specify one metric that a hyperparameter tuning job uses as its objective metric to choose the best training job.

ModelArtifacts
Provides information about the location that is configured for storing model artifacts.

ModelSummary
Provides summary information about a model.

NotebookInstanceLifecycleConfigSummary
Provides a summary of a notebook instance lifecycle configuration.

NotebookInstanceLifecycleHook
Contains the notebook instance lifecycle configuration script.

Each lifecycle configuration script has a limit of 16384 characters.

The value of the $PATH environment variable that is available to both scripts is /sbin:bin:/usr/sbin:/usr/bin.

View CloudWatch Logs for notebook instance lifecycle configurations in log group /aws/sagemaker/NotebookInstances in log stream [notebook-instance-name]/[LifecycleConfigHook].

Lifecycle configuration scripts cannot run for longer than 5 minutes. If a script runs for longer than 5 minutes, it fails and the notebook instance is not created or started.

For information about notebook instance lifestyle configurations, see notebook-lifecycle-config.

NotebookInstanceSummary
Provides summary information for an Amazon SageMaker notebook instance.

ObjectiveStatusCounters
Specifies the number of training jobs that this hyperparameter tuning job launched, categorized by the status of their objective metric. The objective metric status shows whether the final objective metric for the training job has been evaluated by the tuning job and used in the hyperparameter tuning process.

OutputDataConfig
Provides information about how to store model training results (model artifacts).

ParameterRanges
Specifies ranges of integer, continuous, and categorical hyperparameters that a hyperparameter tuning job searches.

ProductionVariant
Identifies a model that you want to host and the resources to deploy for hosting it. If you are deploying multiple models, tell Amazon SageMaker how to distribute traffic among the models by specifying variant weights.

ProductionVariantSummary
Describes weight and capacities for a production variant associated with an endpoint. If you sent a request to the UpdateEndpointWeightsAndCapacities API and the endpoint status is Updating, you get different desired and current values.

ResourceConfig
Describes the resources, including ML compute instances and ML storage volumes, to use for model training.

ResourceLimits
Specifies the maximum number of training jobs and parallel training jobs that a hyperparameter tuning job can launch.

S3DataSource
Describes the S3 data source.

SageMakerClient
A client for the SageMaker API.

SecondaryStatusTransition
Specifies a secondary status the job has transitioned into. It includes a start timestamp and later an end timestamp. The end timestamp is added either after the job transitions to a different secondary status or after the job has ended.

StartNotebookInstanceInput

StopHyperParameterTuningJobRequest

StopNotebookInstanceInput

StopTrainingJobRequest

StopTransformJobRequest

StoppingCondition
Specifies how long model training can run. When model training reaches the limit, Amazon SageMaker ends the training job. Use this API to cap model training cost.

To stop a job, Amazon SageMaker sends the algorithm the SIGTERM signal, which delays job termination for120 seconds. Algorithms might use this 120-second window to save the model artifacts, so the results of training is not lost.

Training algorithms provided by Amazon SageMaker automatically saves the intermediate results of a model training job (it is best effort case, as model might not be ready to save as some stages, for example training just started). This intermediate data is a valid model artifact. You can use it to create a model (CreateModel).

Tag
Describes a tag.

TrainingJobStatusCounters
The numbers of training jobs launched by a hyperparameter tuning job, categorized by status.

TrainingJobSummary
Provides summary information about a training job.

TransformDataSource
Describes the location of the channel data.

TransformInput
Describes the input source of a transform job and the way the transform job consumes it.

TransformJobSummary
Provides a summary information for a transform job. Multiple TransformJobSummary objects are returned as a list after calling ListTransformJobs.

TransformOutput
Describes the results of a transform job output.

TransformResources
Describes the resources, including ML instance types and ML instance count, to use for transform job.

TransformS3DataSource
Describes the S3 data source.

UpdateEndpointInput

UpdateEndpointOutput

UpdateEndpointWeightsAndCapacitiesInput

UpdateEndpointWeightsAndCapacitiesOutput

UpdateNotebookInstanceInput

UpdateNotebookInstanceLifecycleConfigInput

UpdateNotebookInstanceLifecycleConfigOutput

UpdateNotebookInstanceOutput

VpcConfig
Specifies a VPC that your training jobs and hosted models have access to. Control access to and from your training and model containers by configuring the VPC. For more information, see host-vpc and train-vpc.

Enums

AddTagsError
Errors returned by AddTags

CreateEndpointConfigError
Errors returned by CreateEndpointConfig

CreateEndpointError
Errors returned by CreateEndpoint

CreateHyperParameterTuningJobError
Errors returned by CreateHyperParameterTuningJob

CreateModelError
Errors returned by CreateModel

CreateNotebookInstanceError
Errors returned by CreateNotebookInstance

CreateNotebookInstanceLifecycleConfigError
Errors returned by CreateNotebookInstanceLifecycleConfig

CreatePresignedNotebookInstanceUrlError
Errors returned by CreatePresignedNotebookInstanceUrl

CreateTrainingJobError
Errors returned by CreateTrainingJob

CreateTransformJobError
Errors returned by CreateTransformJob

DeleteEndpointConfigError
Errors returned by DeleteEndpointConfig

DeleteEndpointError
Errors returned by DeleteEndpoint

DeleteModelError
Errors returned by DeleteModel

DeleteNotebookInstanceError
Errors returned by DeleteNotebookInstance

DeleteNotebookInstanceLifecycleConfigError
Errors returned by DeleteNotebookInstanceLifecycleConfig

DeleteTagsError
Errors returned by DeleteTags

DescribeEndpointConfigError
Errors returned by DescribeEndpointConfig

DescribeEndpointError
Errors returned by DescribeEndpoint

DescribeHyperParameterTuningJobError
Errors returned by DescribeHyperParameterTuningJob

DescribeModelError
Errors returned by DescribeModel

DescribeNotebookInstanceError
Errors returned by DescribeNotebookInstance

DescribeNotebookInstanceLifecycleConfigError
Errors returned by DescribeNotebookInstanceLifecycleConfig

DescribeTrainingJobError
Errors returned by DescribeTrainingJob

DescribeTransformJobError
Errors returned by DescribeTransformJob

ListEndpointConfigsError
Errors returned by ListEndpointConfigs

ListEndpointsError
Errors returned by ListEndpoints

ListHyperParameterTuningJobsError
Errors returned by ListHyperParameterTuningJobs

ListModelsError
Errors returned by ListModels

ListNotebookInstanceLifecycleConfigsError
Errors returned by ListNotebookInstanceLifecycleConfigs

ListNotebookInstancesError
Errors returned by ListNotebookInstances

ListTagsError
Errors returned by ListTags

ListTrainingJobsError
Errors returned by ListTrainingJobs

ListTrainingJobsForHyperParameterTuningJobError
Errors returned by ListTrainingJobsForHyperParameterTuningJob

ListTransformJobsError
Errors returned by ListTransformJobs

StartNotebookInstanceError
Errors returned by StartNotebookInstance

StopHyperParameterTuningJobError
Errors returned by StopHyperParameterTuningJob

StopNotebookInstanceError
Errors returned by StopNotebookInstance

StopTrainingJobError
Errors returned by StopTrainingJob

StopTransformJobError
Errors returned by StopTransformJob

UpdateEndpointError
Errors returned by UpdateEndpoint

UpdateEndpointWeightsAndCapacitiesError
Errors returned by UpdateEndpointWeightsAndCapacities

UpdateNotebookInstanceError
Errors returned by UpdateNotebookInstance

UpdateNotebookInstanceLifecycleConfigError
Errors returned by UpdateNotebookInstanceLifecycleConfig

Traits

SageMaker
Trait representing the capabilities of the SageMaker API. SageMaker clients implement this trait.